Created release tag for version 1.1 RC4

+++ sventon/dev/java/testsrc/de/berlios/sventon/config/ApplicationConfiguratorTest.java
@@ -52,7 +52,9 @@ public class ApplicationConfiguratorTest extends TestCase {
     } finally {
       IOUtils.closeQuietly(is);
       IOUtils.closeQuietly(os);
-      tempConfigFile.delete();
+      if (tempConfigFile != null) {
+        tempConfigFile.delete();
+      }
     }
 
   }
